removig setup and diagnostics to separate files
