+++ kid.py
@@ -5,6 +5,8 @@ import cffi
 import sys
 import libcloudphxx as libcl
 from libcloudphxx.common import R_v, R_d, c_pd
+from setup import params, opts
+import diagnostics as dg
 import pdb
 
 # CFFI stuff
@@ -17,42 +19,13 @@ ffi.cdef("void save_ptr(char*,void*);")
 
 # Fortran functions (_sp_ means single precision)
 ffi.cdef("void __main_MOD_main_loop();")
-ffi.cdef("void __diagnostics_MOD_save_dg_2d_sp_c(float*, int, int, char*, int,     char*, int,      int   );")
-#                                                field,  nx,  nz,  name,  namelen, units, unitslen, itime
-ffi.cdef("void __diagnostics_MOD_save_bindata_sp_c(float*, int, char*, int,     char*, int      );")
-#                                                  field,  nb,  name,  namelen, units, unitslen
-ffi.cdef("void __diagnostics_MOD_save_dg_2d_bin_sp_c(float*, int, int, int, char*, int,     char*, int,      int   );")
-#                                                    field,  nb,  nx,  nz,  name,  namelen, units, unitslen, itime
 
 # object storing super-droplet model state (to be initialised)
 prtcls = False
 
-# dictionary of simulation parameters
-params = {
-  "real_t" : np.float64,
-  "backend" : libcl.lgrngn.backend_t.serial,
-  "sd_conc" : 128.,
-  "kappa" : .61,
-  "meanr" : .04e-6,
-  "gstdv" : 1.4,
-  "n_tot" : 100e6,
-  "n_bins": 34,             # \__ from the TAU example file @ KiD-A website
-  "bin0_D_upper" : 3.125e-6 # /
-}
-
 arrays = {}
-dt, dx, dz = 0, 0, 0
 first_timestep = True
 last_diag = -1
-opts = libcl.lgrngn.opts_t()
-opts.sstp_cond = 1
-opts.sstp_coal = 1
-opts.cond = True
-opts.coal = True
-opts.adve = True
-opts.sedi = True
-opts.chem = False
-opts.kernel = libcl.lgrngn.kernel_t.geometric
 
 def lognormal(lnr):
   from math import exp, log, sqrt, pi
@@ -75,89 +48,6 @@ def th_dry2kid(th_d, rv):
 
 def rho_kid2dry(rho, rv):
   return rho / (1 + rv) #TODO: I'm assuming that KiD uses rho
-
-def save_helper(arr):
-  # astype() takes keywords arguments for newer numpy versions (1.7?)
-  try: 
-    arr = arr.astype(np.float32, copy=False) 
-  except TypeError:
-    arr = arr.astype(np.float32)
-  arr_ptr = ffi.cast("float*", arr.__array_interface__['data'][0])
-  return arr, arr_ptr
-
-def save_dg(arr, it, name, units):
-  arr, arr_ptr = save_helper(arr)
-  if (arr.ndim == 2):
-    flib.__diagnostics_MOD_save_dg_2d_sp_c(
-      arr_ptr, arr.shape[0], arr.shape[1], 
-      name, len(name), 
-      units, len(units),
-      it
-    )
-  elif (arr.ndim == 3):
-    flib.__diagnostics_MOD_save_dg_2d_bin_sp_c(
-      arr_ptr, arr.shape[0], arr.shape[1], arr.shape[2],
-      name, len(name),
-      units, len(units),
-      it
-    )
-  else:
-    assert(False)
-
-def save_bindata(arr, name, unit):
-  assert(arr.ndim == 1)
-  arr, arr_ptr = save_helper(arr)
-  flib.__diagnostics_MOD_save_bindata_sp_c(
-    arr_ptr, arr.shape[0], 
-    name, len(name), 
-    unit, len(unit)
-  )
-
-def diagnostics(particles, it, size_x, size_z):
-  import math
-
-  tmp = np.empty((size_x-2, size_z))
-  tmp[:,:] = arrays["qv"]
-  save_dg(tmp, it, "aqq", "J")
-
-  # super-droplet concentration per grid cell
-  # TODO: select all particles?
-  particles.diag_sd_conc()
-  save_dg(np.frombuffer(particles.outbuf()).reshape(size_x-2, size_z), it, "sd_conc", "1")
-
-  
-  # temporary arrays (allocating only once)
-  if first_timestep:
-    arrays["mom_0"] = np.empty((params["n_bins"], size_x-2, size_z))
-    arrays["mom_3"] = np.empty((params["n_bins"], size_x-2, size_z))
-    # upper diameter of a bin with values set using mass-doubling scheme
-    arrays["bins_D_upper"] = (2**np.arange(params["n_bins"]))**(1./3) * params["bin0_D_upper"] 
-
-    save_bindata(arrays["bins_D_upper"] * 1e6, "bins_D_upper", "microns")
-    save_bindata((arrays["bins_D_upper"]/2)**3 * libcl.common.rho_w * (4./3) * math.pi, "bins_mass_upper", "kg")
-    save_bindata(np.diff(np.concatenate([np.zeros(1), arrays["bins_D_upper"]])) * 1e6, "dD", "microns")
-
-  # binned wet spectrum
-  r_min = 0.
-  for i in range(params["n_bins"]):
-    # selecting range
-    r_max = arrays["bins_D_upper"][i] / 2
-    particles.diag_wet_rng(r_min, r_max)
-    r_min = r_max
-    # computing 1-st moment
-    particles.diag_wet_mom(0)
-    arrays["mom_0"][i,:,:] = np.frombuffer(particles.outbuf()).reshape(size_x-2, size_z)
-    # computing 3-rd moment
-    particles.diag_wet_mom(3)
-    arrays["mom_3"][i,:,:] = np.frombuffer(particles.outbuf()).reshape(size_x-2, size_z)
-
-  save_dg(arrays["mom_0"], it, "cloud_bin_number", "/kg") 
-  arrays["mom_3"] *= libcl.common.rho_w * (4./3) * math.pi
-  save_dg(arrays["mom_3"], it, "cloud_bin_mass", "kg/kg")
-
-  # binned dry spectrum? - TODO
-  # ...
-  
 
 @ffi.callback("bool(int, float, int, int, double*, double*, double*, double*, double*, double*, double*, double*, double*, double*, double*, double*, double*, double*)")
 def micro_step(it_diag, dt, size_z, size_x, th_ar, qv_ar, rhof_ar, rhoh_ar, 
@@ -214,7 +104,7 @@ def micro_step(it_diag, dt, size_z, size_x, th_ar, qv_ar, rhof_ar, rhoh_ar,
     
     if first_timestep:
       prtcls.init(arrays["thetad"], arrays["qv"], arrays["rhod"], arrays["rhod_Cx"], arrays["rhod_Cz"]) 
-      diagnostics(prtcls, 1, size_x, size_z) # writing down state at t=0
+      dg.diagnostics(prtcls, arrays, 1, size_x, size_z, first_timestep) # writing down state at t=0
 
     # superdroplets: all what have to be done within a timestep
     prtcls.step_sync(opts, arrays["thetad"], arrays["qv"],  arrays["rhod"]) 
@@ -234,7 +124,7 @@ def micro_step(it_diag, dt, size_z, size_x, th_ar, qv_ar, rhof_ar, rhoh_ar,
 
     # diagnostics
     if last_diag < it_diag:
-      diagnostics(prtcls, it_diag, size_x, size_z)
+      dg.diagnostics(prtcls, arrays, it_diag, size_x, size_z, first_timestep)
       last_diag = it_diag
 
     first_timestep = False
